Add a tooltip text column of united variable names and values.

Add a tooltip text column of united variable names and values.

Usage

add_tooltip_text(data, ..., titles = NULL)

Arguments

data

A data frame or tibble.

...

Arguments passed to select (i.e unquoted variables, tidyselect helpers etc). If no arguments provided, uses all columns.

titles

A function to format the variable names, including in rlang lambda format.

Value

A data frame or tibble with a column of text

Examples

iris %>%
  add_tooltip_text() %>%
  head(1)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#>                                                                                                text
#> 1 Sepal.Length: 5.1<br>Sepal.Width: 3.5<br>Petal.Length: 1.4<br>Petal.Width: 0.2<br>Species: setosa

 iris %>%
  add_tooltip_text(Species, tidyselect::contains("Sepal")) %>%
  head(1)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#>                                                       text
#> 1 Species: setosa<br>Sepal.Length: 5.1<br>Sepal.Width: 3.5

  library(snakecase)

 iris %>%
  add_tooltip_text(titles = ~ to_sentence_case(.x)) %>%
  head(1)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#>                                                                                                text
#> 1 Sepal length: 5.1<br>Sepal width: 3.5<br>Petal length: 1.4<br>Petal width: 0.2<br>Species: setosa

 iris %>%
   add_tooltip_text() %>%
   gg_point(x = Sepal.Width,
            y = Sepal.Length,
            col = Species,
            text = text,
            theme = gg_theme("helvetica", x_grid = TRUE, y_grid = TRUE)) %>%
    plotly::ggplotly(tooltip = "text")
#> ℹ {ggblanket} converts unspecified titles using snakecase::to_sentence_case. Use titles = ~.x to leave unspecified titles as is, and/or specify individual titles manually using *_title arguments.
#> ℹ {ggblanket} merges col and fill aesthetics into a single col aesthetic.
#> Warning: Ignoring unknown aesthetics: text
#> Warning: plotly.js does not (yet) support horizontal legend items 
#> You can track progress here: 
#> https://github.com/plotly/plotly.js/issues/53