Quick theme for a ggplot

Quick theme for a ggplot visualisation.

Usage

gg_theme_dark2(
  base_size = 10,
  base_family = "",
  title_size = base_size * 1.1,
  title_pal = "#d3d3d3",
  title_face = "bold",
  title_position = "plot",
  title_hjust = 0,
  title_vjust = 1,
  title_margin = ggplot2::margin(((title_size^0.5) * -0.5) - 3.85, b = title_size * 1.75),
  subtitle_size = other_text_size,
  subtitle_pal = "#a8a8a8",
  subtitle_face = "plain",
  subtitle_hjust = 0,
  subtitle_vjust = 1,
  subtitle_margin = ggplot2::margin(t = subtitle_size * -1.5, b = subtitle_size * 1.75),
  caption_size = other_text_size * 0.9,
  caption_pal = "#777777",
  caption_face = "plain",
  caption_position = "plot",
  caption_hjust = 0,
  caption_vjust = 0,
  caption_margin = ggplot2::margin(t = caption_size * 1.25),
  other_text_size = base_size,
  other_text_pal = "#a8a8a8",
  other_text_face = "plain",
  background_pal_plot = "#0a0a0a",
  background_pal_panel = "#232323",
  background_pal_key = NULL,
  line_linewidth = other_text_size/80,
  line_pal = "#adadad",
  ticks_pal = NULL,
  ticks_linewidth = NULL,
  ticks_length = grid::unit(other_text_size/4, "pt"),
  gridlines_linewidth = other_text_size/100,
  gridlines_pal = "#474747"
)

Arguments

base_size

The base size of the text. Defaults to 10.

base_family

The base font family of the text. Defaults to "".

title_size

The size of the title. Defaults to base_size * 1.1.

title_pal

The colour of the title. Defaults to "#d3d3d3".

title_face

The font face of the title. Defaults to "bold".

title_position

The horizontal alignment of the title (and subtitle) to either "plot" or "panel".

title_hjust

The horizontal adjustment of the title.

title_vjust

The vertical adjustment of the title.

title_margin

The margin of the title. A ggplot2::margin function.

subtitle_size

The size of the subtitle. Defaults to the base_size.

subtitle_pal

The colour of the subtitle. Defaults to "#a8a8a8".

subtitle_face

The font face of the subtitle. Defaults to "plain".

subtitle_hjust

The horizontal adjustment of the subtitle.

subtitle_vjust

The vertical adjustment of the subtitle.

subtitle_margin

The margin of the subtitle. A ggplot2::margin function.

caption_size

The size of the caption. Defaults to base_size * 0.9.

caption_pal

The colour of the caption. Defaults to "#777777".

caption_face

The font face of the caption. Defaults to "plain".

caption_position

The horizontal alignment of the caption to either "plot" or "panel".

caption_hjust

The horizontal adjustment of the caption.

caption_vjust

The vertical adjustment of the caption.

caption_margin

The margin of the caption. A ggplot2::margin function.

other_text_size

The size of all text other than the title, subtitle and caption. Defaults to the base_size.

other_text_pal

The colour of all text other than the title, subtitle or caption. Defaults to "#d3d3d3".

other_text_face

The font face of all text other than the title, subtitle or caption. Defaults to "plain".

background_pal_plot

The colour of the plot background colour. Defaults to "#000000".

background_pal_panel

The colour of the panel background colour. Defaults to "#232323".

background_pal_key

The colour of the legend key. Defaults to the background_pal_panel.

line_linewidth

The linewidth of the axis line. Defaults to the other_text_size / 80 (i.e. 0.125).

line_pal

The colour of the axis line. Defaults to "#adadad".

ticks_pal

The colour of the ticks. Defaults to the line_pal.

ticks_linewidth

The linewidth of the axis ticks. Defaults to the line_linewidth.

ticks_length

The length of the ticks. A grid::unit function.

gridlines_linewidth

The linewidth of the vertical major gridlines. Defaults to other_text_size / 100 (i.e. 0.1).

gridlines_pal

The colour of the vertical major gridlines. Defaults to "#474747".

Value

A ggplot theme.